[Build] Fix all compilation warnings

+++ Logic/source/AI/WaveManager.cpp
@@ -26,7 +26,7 @@ void WaveManager::loadFile()
     m_waveInformation.timeMul   = waveInfo.floats["timeMul"];
     m_waveInformation.timeBonus = waveInfo.floats["timeBonus"];
     m_waveInformation.bossWaves = splitInts(waveInfo.strings["bossWaves"], FILE_DELIM);
-    m_waveInformation.bossTimeBonus = waveInfo.floats["bossTimeBonus"];
+    m_waveInformation.bossTimeBonus = (int)waveInfo.floats["bossTimeBonus"];
 }
 
 WaveManager::EntitiesInWave WaveManager::getEntities(int waveId) const
